Put non-authoritative mp data parsing into separate file.
