+++ right_to_ask_api/src/lib.rs
@@ -15,6 +15,7 @@ pub mod common_file;
 pub mod censorship;
 pub mod minister;
 mod mp_non_authoritative;
+mod parse_non_authoritative_mp_data;
 
 #[cfg(test)]
 mod tests {
